requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
requests
pandas
google-cloud-aiplatform
openai
dotenv